I rewrote the main function of Burgers_Same
My goal is to essentially recreate the results of this paper. To do that,
I need the data they used. Thus, I'm trying to understand their data
loading code (the first bit of the main function).

This is some weird code. To begin, it has very few comments, so I'm
having to parse it line-by-line (which is taking a while). Next,
the original code loaded in from file twice, even though this was
totally unnecessary. (see my comments/changes). He also randomly
reordered the spatial points to enforce the BC at, as well as the
temporal points to enforce the Periodic BCs at. I have no clue why he
did this, but okay. Honestly, I'm confused about the rationale behind
most of the code in the "##### Points to learn the solution. #####"
subsection.

Either way, I think I got what I was looking for, which is good.